+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/multi/MatrixMultiMessenger.kt
@@ -82,6 +82,7 @@ class MatrixMultiMessengerImpl private constructor(
         }
         activeMatrixMessenger.value?.close()
         di.get<MatrixMultiMessengerConfiguration>().httpClientEngine?.close()
+        di.close()
     }
 }